requirements.txt
openai
# langchain
streamlit
langchain-community
langchain-openai